Updated projects/IT_RoadSigns_120classes.

+++ YOLOMixedSizeDatasetCreator.py
@@ -13,17 +13,16 @@
 # limitations under the License.
 #
 # 
-# 2022/05/10 copyright (c) antillia.com
-# 2022/05/31 Added YOLO2COCOConverter to create a ground truth json(coco_annotation.json) 
-# from yolo annotation files in outout_dir
 
-# YOLOTestDatasetCreator.py
+
+# YOLOMixedSizeDatasetCreator.py
 
 # importing libraries
 import os
 import sys
 from PIL import Image, ImageDraw, ImageFilter
 import shutil
+import uuid
 
 import numpy as np
 import glob
@@ -37,8 +36,8 @@ from YOLO2COCOConverter import YOLO2COCOConverter
 
 class YOLOMixedSizeDatasetCreator:
   # Constructor
-  def __init__(self, dataset_name, background_size, max_image_size, classes_file, 
-              num_images_per_background=5):
+  def __init__(self, dataset_name, background_size, max_image_size, classes_file, debug):
+  
     self.dataset_prefix = dataset_name + "_"
     self.labelmap = None
     self.classes = []
@@ -59,8 +58,10 @@ class YOLOMixedSizeDatasetCreator:
     self.BACKGROUND_IMAGE_HEIGHT= background_size[1]
 
     self.MAX_IMAGE_WIDTH        = max_image_size[0]
-    self.NUM_IMAGES_PER_BACKGROUND = num_images_per_background
+    self.MAX_IMAGE_HEIGHT       = max_image_size[1]
 
+    self.debug = debug
+    
   def getClassIndex(self, sname):
     index = -1
     for i, name in enumerate(self.classes):
@@ -71,7 +72,6 @@ class YOLOMixedSizeDatasetCreator:
          break
     return index
 
-  # 2022/05/27 modified
   def getClassName(self, sname):
     cname = None
     pos = sname.find("___")
@@ -121,56 +121,75 @@ class YOLOMixedSizeDatasetCreator:
     return rc
 
   def create(self, backgrounds_dir, images_dir, output_dir, scalings):
+    if os.path.exists(backgrounds_dir) == False:
+       raise Exception("Not found backgrounds_dir {}".format(backgrounds_dir))
+    if os.path.exists(images_dir) == False:
+       raise Exception("Not found images_dir {}".format(images_dir))
+    
+    if os.path.exists(output_dir):
+       shutil.rmtree(output_dir) 
+
+    if os.path.exists(output_dir) == False:
+       os.makedirs(output_dir)  
+    
     background_pattern  = backgrounds_dir  + "/*.jpg"
     background_files    = glob.glob(background_pattern)
     images_pattern      = images_dir + "/*.png"
-    image_files        = glob.glob(images_pattern)
+    image_files         = glob.glob(images_pattern)
 
     for image_file in image_files:
+
       for scaling in scalings:
 
         [background_file] = random.sample(background_files, 1)
 
         background = Image.open(background_file)
-        #samples = random.sample(images_files, self.NUM_IMAGES_PER_BACKGROUND)
         
         try:      
-          print("------ {} background {}".format(n, background))
-          fname = self.dataset_prefix + str(1000+n) + ".jpg"
-          aname = self.dataset_prefix + str(1000+n) + ".txt"
+          uuid4= uuid.uuid4()
+
+          fname = str(uuid4) + ".jpg"
+          aname = str(uuid4) + ".txt"
           outputfile = os.path.join(output_dir, fname)
           annotation = os.path.join(output_dir, aname) 
           print("=== output {}".format(outputfile))
           i = 0
           WIDTH = self.MAX_IMAGE_WIDTH #240
-          #w = 240
-          g = int((self.BACKGROUND_IMAGE_WIDTH -WIDTH)/2)
-          X_POS = 10  # 20  2022/06/27
-          Y_POS = 10  # 40  2022/06/27
-          X_W   = 245 # 210 2022/06/27
-          Y_H   = 110 # 120 2022/06/27
+ 
+          X_POS = 10  
+          Y_POS = 10  
 
           SPACE = " "
           NL    = "\n"
           bgwidth  = self.BACKGROUND_IMAGE_WIDTH
           bgheight = self.BACKGROUND_IMAGE_HEIGHT
           with open(annotation, "w") as f:
-            for i, sample in enumerate(samples):
-              print(" === i {} sample {}".format(i, sample))
-              image = Image.open(sample).convert("RGBA")
+              image = Image.open(image_file).convert("RGBA")
               if image == None:
-                raise Exception("Failed to open an image file " + sample)
+                raise Exception("Failed to open an image file " + image_file)
+
+              ow, oh = image.size
+              rw = int(ow * scaling)
+              rh = int(oh * scaling)
+              image = image.resize((rw, rh))
               W, H = image.size
               print("--- W: {} H: {}".format(W,H))
-              px   = X_POS + X_W*i
-              py   = Y_POS + Y_H*i        
-              sname = os.path.basename(sample)
+
+              if W> self.MAX_IMAGE_WIDTH or H > self.MAX_IMAGE_HEIGHT:
+                print("--------------skipping image file {} scaling {}".format(image_file, scaling))
+                
+                continue
+
+              print("--- W: {} H: {}".format(W,H))
+              px   = X_POS
+              py   = Y_POS        
+              sname = os.path.basename(image_file)
               sname = self.getClassName(sname)
               classIndex = self.getClassIndex(sname)
               if classIndex == -1:
                 print("--- Not found class {}".format(sname))
                 raise Exception("Invalid index----------")
-
+            
               #YOLO annotation
               centerx = (px + W/2 ) / float(bgwidth)
               centery = (py + H/2 ) / float(bgheight)
@@ -182,22 +201,29 @@ class YOLOMixedSizeDatasetCreator:
               width   = round(width,   4)
               height  = round(height,  4)
 
-              ann     =  str(classIndex) + SPACE + str(centerx) + SPACE + str(centery) + SPACE + str(width) + SPACE + str(height) + NL
-              f.write(ann)
-              self.paste(background, image, px, py)
-              image.close()
+              line    =  str(classIndex) + SPACE + str(centerx) + SPACE + str(centery) + SPACE + str(width) + SPACE + str(height) + NL
+              f.write(line)
 
-              print("-----pasted {}".format(sample))
-              
+              print("--- annotation file:{} annotation:{}".format(annotation, line))
+  
+              self.paste(background, image, px, py)
+              if self.debug:
+                draw = ImageDraw.Draw(background) 
+                rectcolor = (255, 0, 0) 
+                linewidth = 2 
+                draw.rectangle([(px, py), (px+W, px+H)], \
+                      outline=rectcolor, width=linewidth) 
+
+  
           print("=== saved outputfile {}".format(outputfile))
           background.save(outputfile, quality=95)
-          background.close()
+
 
         except:
           traceback.print_exc()
-          #raise Exception("Invalid index")
 
-# python YOLOTestDatasetCreator.py  ./projects/IT-RoadSigns-120classes/configs/yolo_mixed_dataset_creator.conf
+# python YOLOTestDatasetCreator.py  ./projects/IT-RoadSigns-120classes/configs/yolo_mixed_size_dataset_creator.conf
+
 if __name__ == "__main__":
   config_ini  = ""
 
@@ -210,39 +236,28 @@ if __name__ == "__main__":
     print("--- config_ini {}".format(config_ini))
     parser = ConfigParser(config_ini)
     DATASET = "dataset"
-    target  = "mixed_size"
+    TRAIN   = "train"
+    VALID   = "valid"
     dataset_name    = parser.get(DATASET, "name")
     background_size = parser.get(DATASET, "background_size")
     max_image_size  = parser.get(DATASET, "max_image_size")
     classes_file    = parser.get(DATASET, "classes")
-
-    backgrounds_dir = parser.get(target,  "backgrounds_dir")
-    images_dir      = parser.get(target,  "images_dir")
-    output_dir      = parser.get(target,  "output_dir")
-    scalings        = parser.get(target,  "scalings")
-    if os.path.exists(backgrounds_dir) == False:
-       raise Exception("Not found backgrounds_dir {}".format(backgrounds_dir))
-    if os.path.exists(images_dir) == False:
-       raise Exception("Not found images_dir {}".format(images_dir))
-    
-    if os.path.exists(output_dir):
-       shutil.rmtree(output_dir) 
-
-    if os.path.exists(output_dir) == False:
-       os.makedirs(output_dir)  
-
+    debug           = parser.get(DATASET, "debug")
+  
     if os.path.exists(classes_file) == False:
        raise Exception("Not found classes_file {}".format(classes_file))
+    targets = [TRAIN, VALID]
+      
+    creator = YOLOMixedSizeDatasetCreator(dataset_name, background_size, max_image_size, classes_file, debug)
 
-    creator = YOLOMixedSizeDatasetCreator(dataset_name, background_size, max_image_size, classes_file)
-    creator.create(backgrounds_dir, images_dir, output_dir, scalings)
+    for target in targets:
+      backgrounds_dir = parser.get(target,  "backgrounds_dir")
+      images_dir      = parser.get(target,  "images_dir")
+      output_dir      = parser.get(target,  "output_dir")
+      scalings        = parser.get(target,  "scalings")
+ 
+      creator.create(backgrounds_dir, images_dir, output_dir, scalings)
     
-    #2022/05/31
-    #Create a coco_annotation.json(ground truth json) from yolo annoation files in outout_dir
-    #coco_json    = os.path.join(output_dir, "annotation.json")
-
-    #converter = YOLO2COCOConverter(classes_file)  
-    #converter.run(output_dir, output_dir, coco_json)
 
   except:
     traceback.print_exc()
